Externalize css; Use dvh instead of js hack
Because I want to reference the css from a different domain, rather than
replicate it when I need it, I've moved it out to a file.

Since I was in there, I went ahead and got rid of the vh hack now that
I can use the `dvh` unit.

Also removed the .js class since I wasn't using it.